requirements.txt
customtkinter==5.2.2
pillow==10.2.0
requests==2.31.0 